requirements.txt
blinker==1.9.0
capstone==5.0.6
click==8.3.1
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psutil==7.2.1
pyelftools==0.32
Werkzeug==3.1.4
